Introduce IOExecTemplate operator
This operator makes it a bit easier to invoke processes by using the `%s`
format specifier as an argument placeholder.

Note that this this does not allow for spaces in the process path (e.g.
`C:\Program Files\...`) and only supports a single `%s` within a word
(space delimited part of the template).