Defer evaluation of expression to some later time or never.
For TLC's simulation mode, later is defined as the point of time when
simulation has chosen a successor state to extend the prefix of the
current behavior.

Use case: TLCDefer(TLCSet(42, someVal)) evals TLCSet(42, someVal) only
for the states of the behavior and not the set of all successor states
of all states in the behavior. In other words, someVal is only set for
the states of the behavior.

[Feature]