Use gson instead of the jackson library for json operations
It replaces the jackson dependencies by the gson one so we have
a compatible license (see issue link above).

It's a very direct conversion, I've even maintained method names like
`getArrayNode` or `getObjectNode`, it does not appear to be worth to
change it to follow the implementation classes names.

Fixes Github issue #29
https://github.com/tlaplus/CommunityModules/issues/29

[Refactor]