Add a type test in MapThenFoldSet
I ran into an issue with TLC and folds several times. If somebody passes an integer instead of a set for the parameter `S`, TLC produces an unreadable error for the definition of `iter[s \in SUBSET S]`. There is a simple fix for that: We can compare `S` to the empty set. If someone passed a non-set, TLC would complain and produce a nice error trace.